Refactored both functions and resolved all errors
The merge function has been almost completely overhauled as the previous
version was still a bit too close to the Java example that I was trying
to base the function off of. The mergeSort function got the same kind
of treatment but the changes there were far smaller.

The empty 'data' array was removed along with the 'start' and 'end'
variable declarations. The 'tempArray' was changed back to an empty
array from the Array constructor that was there before (Java example).
There are no more 'start, mid, end' parameters; just simple 'left' and
'right' parameters representing split arrays themselves, not indexes
within arrays fed to the merge function. The while loops were condensed
down to 1 that deals with the actual merging of the sub-arrays while
both still have values. Any remaining values left over on either array
when the other is empty is dealt with in the 'return' statement using
the 'concat' method. We also have the two indexes for each array
declared as 0 at the top, simplifying from the 3 we had before (one was
a 'tracker' of sorts for the tempArray). Deleted the 'for' loop as we
no longer needed it for 'copying' elements to another array.

The mergeSort basically stayed the same. The only major difference was
using the 'slice' method for the actual splitting of the arrays instead
of declaring multiple parameters in a conditional. The only conditional
now is the base case. We also now use 'Math.floor' to calculate the
midpoint of any array instead of trying to add two parameters together
and dividing by 2.